적 체력바를 화면에 띄우고 테스트해라
체력바를 띄우기 이전에 체력을 그동안 int32로 관리하고있었는데,
이보다는 확장성을 위해서 float으로 변경하였다.
(예를 들면 방어력 감소 %와 같은 것을 이용하려면 float이어야 하기 때문)

Enemy 클래스에 Widget Component를 부착하고 이 Widget이 보여질때만
Billboard효과를 주었다. (Tick 함수)
또한 OverlapEvent가 발생할 시, 체력을 20씩 감소하는 효과를 통해서 테스트하였다.

Projectile도 Overlap Event가 발생하게 된다면 그 즉시 삭제하는 코드를 생성하였다.

+++ Source/DefenseGame/CharacterProgressBarWidget.h
@@ -24,7 +24,7 @@ protected:
 
 public:
 	UFUNCTION()
-	void OnChangeHP(int32 CurrentHP, int32 MaxHP);
+	void OnChangeHP(float CurrentHP, float MaxHP);
 
 private:
 	UPROPERTY(EditDefaultsOnly, meta = (BindWidget), meta = (AllowPrivateAccess = "true"))
